+++ src/identifier.rs
@@ -498,7 +498,7 @@ mod tests {
 #[cfg(feature = "serde")]
 #[cfg_attr(docsrs, doc(cfg(feature = "serde")))]
 mod serde_support {
-    use super::{fmt, Scru128Id};
+    use super::{fmt, str, Scru128Id};
     use serde::{de, Deserializer, Serializer};
 
     impl serde::Serialize for Scru128Id {
@@ -535,9 +535,13 @@ mod serde_support {
         }
 
         fn visit_bytes<E: de::Error>(self, value: &[u8]) -> Result<Self::Value, E> {
-            <[u8; 16]>::try_from(value)
-                .map(Self::Value::from)
-                .map_err(de::Error::custom)
+            match <[u8; 16]>::try_from(value) {
+                Ok(array_value) => Ok(Self::Value::from(array_value)),
+                Err(err) => match str::from_utf8(value) {
+                    Ok(str_value) => self.visit_str(str_value),
+                    _ => Err(de::Error::custom(err)),
+                },
+            }
         }
     }
 
@@ -608,6 +612,10 @@ mod serde_support {
                 // deserialize the other format regardless of human-readability configuration
                 serde_test::assert_de_tokens(&e.readable(), &[Token::Bytes(bytes)]);
                 serde_test::assert_de_tokens(&e.compact(), &[Token::Str(text)]);
+
+                // deserialize textual representation even if passed as byte slice
+                serde_test::assert_de_tokens(&e.readable(), &[Token::Bytes(text.as_bytes())]);
+                serde_test::assert_de_tokens(&e.compact(), &[Token::Bytes(text.as_bytes())]);
             }
         }
     }